WIP:feature38-DAOの定義:In-Memory Databaseを使ったテストの作成

+++ app/src/main/java/com/example/routinetimerclone/data/entitiy/TaskEntity.kt
@@ -17,8 +17,8 @@ import androidx.room.PrimaryKey
     indices = [androidx.room.Index("parentRoutineId")],
 )
 data class TaskEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     val name: String,
     val seconds: Int,
-    val parentRoutineId: Int,
+    val parentRoutineId: Long,
 )
