+++ app/src/main/java/com/example/routinetimerclone/data/entitiy/mapper/TaskModelMapper.kt
@@ -12,7 +12,7 @@ class TaskModelMapper {
 
         fun toEntity(
             task: Task,
-            parentRoutineId: Int,
+            parentRoutineId: Long,
         ): TaskEntity {
             return TaskEntity(task.id, task.name, task.duration.getTotalSeconds(), parentRoutineId)
         }
